[REFACTOR] rearrange file structure
