--- src/actions/FightEvaluation.Actions.ts
@@ -1,6 +1,0 @@
-export interface IFightEvaluationProps{
-	rankExercise: any;
-	rankCriterias: any[];
-    performer: any;
-    exercise: any;
-}
