--- src/actions/PhysicalEvaluation.Actions.ts
@@ -1,5 +1,0 @@
-export interface IPhysicalEvaluationProps{
-    rankExercise: any;
-    performer: any;
-    exercise: any;
-}
