--- src/actions/TaoluEvaluation.Actions.ts
@@ -1,6 +1,0 @@
-export interface ITaoluEvaluationProps{
-    rankExercise: any;
-    rankCriterias: any[];
-    performer: any;
-    exercise: any;
-}
