+++ src/actions/evaluate/EvaluateFight.Actions.ts
@@ -1,0 +1,6 @@
+export interface IEvaluateFightProps{
+	rankExercise: any;
+	rankCriterias: any[];
+    performer: any;
+    exercise: any;
+}
