+++ src/actions/evaluate/EvaluatePhysical.Actions.ts
@@ -1,0 +1,5 @@
+export interface IEvaluatePhysicalProps{
+    rankExercise: any;
+    performer: any;
+    exercise: any;
+}
