+++ src/actions/evaluate/EvaluateTaolu.Actions.ts
@@ -1,0 +1,6 @@
+export interface IEvaluateTaoluProps{
+    rankExercise: any;
+    rankCriterias: any[];
+    performer: any;
+    exercise: any;
+}
