+++ src/custom/components/SearchInput/SearchInput.css
@@ -1,0 +1,7 @@
+.SearchInput {
+	display: inline-block;
+}
+
+.SearchInput .ant-input-affix-wrapper .ant-input-suffix {
+	right: 0;
+}
