+++ src/custom/components/SearchInput/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './SearchInput';
