+++ src/custom/index.tsx
@@ -1,0 +1,1 @@
+export {default as SearchInput} from './components/SearchInput';
