+++ src/evaluate/components/EvaluateExercise/EvaluateCriterias/EvaluateCriterias.css
@@ -1,0 +1,13 @@
+.EvaluateCriterias .criteria {
+    display: flex;
+    align-items: center;
+    margin: 16px 32px;
+}
+
+.EvaluateCriterias .criteria-name {
+    margin-left: 24px;
+}
+
+.EvaluateCriterias .ant-select {
+    width: 64px;
+}
