+++ src/evaluate/components/EvaluateExercise/EvaluateCriterias/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './EvaluateCriterias';
