+++ src/evaluate/components/EvaluateExercise/EvaluateExercise.css
@@ -1,0 +1,39 @@
+.EvaluateExercise{
+    display: flex;
+    flex-direction: column;
+    flex-grow: 2;
+}
+
+.EvaluateExercise .header {
+    display: flex;
+    align-items: center;
+    border-bottom: 1px solid rgba(0,0,0,.12);
+    padding: 16px;
+    flex-shrink: 0;
+}
+
+.EvaluateExercise .name {
+    font-weight: 600;
+}
+
+.EvaluateExercise .type {
+    margin-left: auto;
+    font-size: 80%;
+}
+
+.EvaluateExercise .body {
+    padding: 16px;
+    overflow-y: auto;
+    display: flex;
+    flex-direction: column;
+}
+
+.EvaluateExercise .description {
+    margin-bottom: 16px;
+}
+
+.EvaluateExercise .image{
+    margin-left: auto;
+    margin-right: auto;
+    margin-bottom: 32px;
+}
