+++ src/evaluate/components/EvaluateExercise/EvaluateFight/Timer/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './Timer';
