+++ src/evaluate/components/EvaluateExercise/EvaluateFight/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './EvaluateFight';
