+++ src/evaluate/components/EvaluateExercise/EvaluatePhysical/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './EvaluatePhysical';
