+++ src/evaluate/components/EvaluateExercise/EvaluateTaolu/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './EvaluateTaolu';
