+++ src/evaluate/components/EvaluateGroup/EvaluateGroup.css
@@ -1,0 +1,44 @@
+.EvaluateGroup {
+    display: flex;
+    flex-grow: 2;
+    -webkit-box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
+    box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
+    border-radius: 8px;
+    background-color: white;
+}
+
+.EvaluateGroup > div {
+    display: flex;
+    flex-direction: column;
+    position: relative;
+}
+
+.EvaluateGroup .ant-menu {
+    overflow-y: auto;
+    overflow-x: hidden;
+    background-color: transparent;
+}
+
+.EvaluateGroup .list-header {
+    padding: 16px;
+    font-weight: 600;
+    border-bottom: 1px solid rgba(0,0,0,.12);
+}
+
+.EvaluateGroup .list {
+    border-right: 1px solid rgba(0,0,0,.12);
+}
+
+.EvaluateGroup .exercise-panel {
+    display: flex;
+    flex-direction: column;
+}
+
+.EvaluateGroup .next {
+    border-top: 1px solid rgba(0,0,0,.12);
+    padding: 16px;
+}
+
+.EvaluateGroup .next button {
+    float: right;
+}
