+++ src/evaluate/components/EvaluateGroup/index.tsx
@@ -1,0 +1,1 @@
+export {default as EvaluateGroup} from './EvaluateGroup';
