+++ src/evaluate/index.tsx
@@ -1,0 +1,1 @@
+export {default as EvaluateGroup} from './components/EvaluateGroup/EvaluateGroup';
