+++ src/evaluations/components/EvaluationsList/EvaluationsList.css
@@ -1,0 +1,94 @@
+.EvaluationsList {
+    display: flex;
+    flex-grow: 2;
+    -webkit-box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
+    box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
+    border-radius: 8px;
+    background-color: white;
+    flex-direction: column;
+}
+
+.EvaluationsList .header {
+    border-bottom: 1px solid rgba(0,0,0,.12);
+    padding: 16px;
+}
+
+.EvaluationsList .top{
+    display: flex;
+}
+
+.EvaluationsList .top button {
+    margin-left: auto;
+}
+
+.EvaluationsList .evaluations {
+    overflow-y: auto;
+}
+
+.EvaluationsList .evaluation  {
+    -webkit-box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
+    box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
+    border-radius: 8px;
+    background-color: white;
+    flex-shrink: 0;
+    margin: 16px;
+    width: auto;
+    padding: 16px;
+    cursor: pointer;
+    display: flex;
+    flex-direction: column;
+    position: relative;
+}
+
+.EvaluationsList .evaluation button{
+    display: flex;
+    margin-left: inherit;
+}
+
+.EvaluationsList .SearchInput {
+    width: 400px;
+    margin-left: 16px;
+}
+
+.EvaluationsList .select {
+    width: 200px;
+}
+
+.EvaluationsList .evaluation-header {
+    display: flex;
+    margin-bottom: 16px;
+}
+
+.EvaluationsList .title-type {
+    display: flex;
+    flex-direction: column;
+}
+
+.EvaluationsList .title {
+    font-size: 130%;
+    font-weight: 700;
+}
+
+.EvaluationsList .type {
+    font-size: 80%;
+}
+
+.EvaluationsList .actions,
+.EvaluationsList .date {
+    margin-left: auto;
+}
+
+.EvaluationsList .body,
+.EvaluationsList .actions {
+    display: flex;
+    align-items: flex-end;
+}
+
+.EvaluationsList .descriptions {
+    display: flex;
+    flex-direction: column;
+}
+
+.EvaluationsList .actions button {
+    margin-left: 16px;
+}
