--- src/evaluations/styles/CriteriasEvaluation.css
@@ -1,13 +1,0 @@
-.CriteriasEvaluation .criteria {
-    display: flex;
-    align-items: center;
-    margin: 16px 32px;
-}
-
-.CriteriasEvaluation .criteria-name {
-    margin-left: 24px;
-}
-
-.CriteriasEvaluation .ant-select {
-    width: 64px;
-}
