--- src/evaluations/styles/ExerciseEvaluation.css
@@ -1,39 +1,0 @@
-.ExerciseEvaluation{
-    display: flex;
-    flex-direction: column;
-    flex-grow: 2;
-}
-
-.ExerciseEvaluation .header {
-    display: flex;
-    align-items: center;
-    border-bottom: 1px solid rgba(0,0,0,.12);
-    padding: 16px;
-    flex-shrink: 0;
-}
-
-.ExerciseEvaluation .name {
-    font-weight: 600;
-}
-
-.ExerciseEvaluation .type {
-    margin-left: auto;
-    font-size: 80%;
-}
-
-.ExerciseEvaluation .body {
-    padding: 16px;
-    overflow-y: auto;
-    display: flex;
-    flex-direction: column;
-}
-
-.ExerciseEvaluation .description {
-    margin-bottom: 16px;
-}
-
-.ExerciseEvaluation .image{
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: 32px;
-}
