--- src/evaluations/styles/FightEvaluation.css
@@ -1,19 +1,0 @@
-.FightEvaluation .ant-table-title
-{
-  text-align: center;
-}
-
-.FightEvaluation .Score td
-{
-    display: flex;
-}
-
-.FightEvaluation .ant-input{
-    width: 20px;
-}
-
-.FightEvaluation .Buttons{
-    display: flex;
-    justify-content : space-around;
-    bottom:20%; 
-}
