--- src/evaluations/styles/GroupEvaluation.css
@@ -1,44 +1,0 @@
-.GroupEvaluation {
-    display: flex;
-    flex-grow: 2;
-    -webkit-box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
-    box-shadow: 0 1px 2px 0 rgba(60,64,67,.3), 0 1px 3px 1px rgba(60,64,67,.15);
-    border-radius: 8px;
-    background-color: white;
-}
-
-.GroupEvaluation > div {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-}
-
-.GroupEvaluation .ant-menu {
-    overflow-y: auto;
-    overflow-x: hidden;
-    background-color: transparent;
-}
-
-.GroupEvaluation .list-header {
-    padding: 16px;
-    font-weight: 600;
-    border-bottom: 1px solid rgba(0,0,0,.12);
-}
-
-.GroupEvaluation .list {
-    border-right: 1px solid rgba(0,0,0,.12);
-}
-
-.GroupEvaluation .exercise-panel {
-    display: flex;
-    flex-direction: column;
-}
-
-.GroupEvaluation .next {
-    border-top: 1px solid rgba(0,0,0,.12);
-    padding: 16px;
-}
-
-.GroupEvaluation .next button {
-    float: right;
-}
