--- src/visualisationevaluation/index.tsx
@@ -1,1 +1,0 @@
-export * from './components/VisualisationEvaluation';
