--- src/visualisationevaluation/styles/Search.css
@@ -1,7 +1,0 @@
-.Search {
-	display: inline-block;
-}
-
-.Search .ant-input-affix-wrapper .ant-input-suffix {
-	right: 0;
-}
